% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.GCElm.r
\name{confint.GCElm}
\alias{confint.GCElm}
\title{Confidence Intervals for Model Parameters}
\usage{
\method{confint}{GCElm}(object, parm, level = 0.95, ...)
}
\arguments{
\item{object}{an object of class \code{lm}.}

\item{parm}{a specification of which parameters are to be given
confidence intervals, either a vector of numbers or a vector
of names. If missing, all parameters are considered.}

\item{level}{the confidence level required.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A matrix with columns giving lower and upper confidence limits
for each parameter.
}
\description{
Computes confidence intervals for one or more parameters in a fitted
model. This is a method for objects of class \code{lm}.
}
\details{
The intervals are based on the asymptotic normality of the estimates
and use the estimated covariance matrix of the coefficients.
}
\examples{
fit <- lm(Sepal.Length ~ Sepal.Width + Petal.Length, data = iris)
confint(fit)

}
\seealso{
\code{\link[stats]{coef}}, \code{\link[GCEmodels]{GCElm}}
}
