% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCElm.fit.r
\name{GCElm.fit}
\alias{GCElm.fit}
\title{Generalized Cross Entropy Linear Regression Models}
\usage{
GCElm.fit(y, X, Z, v, nu, p0, w0, k.sigma = 3, control = GCElm.control())
}
\arguments{
\item{y}{numeric, n (Nx1) vector representing the dependent variable where N is the number of observations.}

\item{X}{numeric, n (NxK) matrix representing a set of independent variables where K is number of regressors.}

\item{Z}{numeric, An (KxM) matrix representing support spaces the for regression coefficients where M is the dimension of the support spaces.}

\item{v}{An optional argument representing a support space for error terms: (a) if missing then v is a (5x1) vector of equally spaced points in [a,b] interval; (b) if a scalar (e.g. H) then v is a (Hx1) vector of equally spaced points in [a,b] interval; (c) can be a user-supplied vector; (d) can be a user-supplied matrix. Please note that in case (a) and (b) the [a,b] interval is centered around zero, and a and b are calculated using the empirical three-sigma rule Pukelsheim (1994).}

\item{nu}{numeric, optional: A weight parameter representing the trade-off between prediction and precision.}

\item{p0}{numeric, optional: Prior probabilities associated with the regression coefficients.}

\item{w0}{numeric, optional: Prior probabilities associated with the error terms.}

\item{k.sigma}{Implement the k-sigma rule}

\item{control}{list, a list of parameters for controlling the fitting process; for GCElm.fit this is passed to \code{\link{GCElm.control}}.}
}
\value{
A \code{list} with the following elements:

* \code{lambda}, estimated lagrange multipliers

* \code{beta}, regression coefficients

* \code{var_beta}, variance-covariance matrix of the regression coefficients

* \code{p}, estimated probabilities associated with the regressions coefficients

* \code{w}, estimated probabilities associated with the error terms

* \code{e}, estimated residuals

* \code{Sp}, the (signal) information of the whole system

* \code{Sp_k}, the (signal) information associated with the k-th regression coefficient

* \code{H_p_w}, value of the joint entropies of p and w at the final iteration

* \code{dH}, delta-H from the Entropy Concentration Theorem

* \code{ER}, entropy-ratio statistic

* \code{Pseudo-R2}, pseudo R-squared

* \code{conv}, convergence (same as in the \code{lbfgs} function)
}
\description{
Fitting generalized cross entropy (GCE) linear models
}
\details{
Mettere qui eventuali details.
}
\examples{
set.seed(1234)
N <- 5000
K <- 10
y <- runif(N)
X <- matrix(runif(N*K), nrow = N, ncol = K)
X <- cbind(rep(1, N), X)
Z <- matrix(rep(c(-100, 50, 0, 50, 100), K+1), nrow = K+1, byrow = TRUE)
GCEfit <- GCElm.fit(y, X, Z)
data.frame(beta = GCEfit$beta,
           beta_lb = GCEfit$beta-1.96*sqrt(diag(GCEfit$var_beta)),
           beta_ub = GCEfit$beta+1.96*sqrt(diag(GCEfit$var_beta))
           )
}
\references{
Golan (1996)
}
\author{
Marco Sandri, Enrico Ciavolino, Maurizio Carpita (\email{gcemodels@gmail.com})
}
