\name{GCEmodels-package}
\alias{GCEmodels-package}
\alias{GCEmodels}
\docType{package}
\title{
\packageTitle{GCEmodels}
}
\description{
\packageDescription{GCEmodels}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{GCEmodels}
\packageIndices{GCEmodels}
An overview of how to use the package, including the most important functions
}
\author{
\packageAuthor{GCEmodels}

Maintainer: \packageMaintainer{GCEmodels}
}
\references{
Literature or other references for background information
}

\keyword{ package }
\seealso{
Optional links to other man pages, e.g. 
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.   
  }
}
