% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.GCElogit.R
\name{print.summary.GCElogit}
\alias{print.summary.GCElogit}
\title{Print Method for \code{summary.GCElogit}}
\usage{
\method{print}{summary.GCElogit}(x, digits = max(3L, getOption("digits") -
  3L), ...)
}
\arguments{
\item{x}{An object of class \code{"summary.GCElogit"}.}

\item{digits}{Number of significant digits to use when printing.}

\item{...}{Further arguments passed to \code{\link[stats]{printCoefmat}}.}
}
\value{
The input object \code{x}, invisibly.
}
\description{
Prints a \code{summary.GCElogit} object in a \code{summary.glm}-like format.
}
