\name{ProvaProjEigen-package}
\alias{ProvaProjEigen-package}
\alias{ProvaProjEigen}
\docType{package}
\title{
\packageTitle{ProvaProjEigen}
}
\description{
\packageDescription{ProvaProjEigen}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ProvaProjEigen}
\packageIndices{ProvaProjEigen}
An overview of how to use the package, including the most important functions
}
\author{
\packageAuthor{ProvaProjEigen}

Maintainer: \packageMaintainer{ProvaProjEigen}
}
\references{
Literature or other references for background information
}

\keyword{ package }
\seealso{
Optional links to other man pages, e.g. 
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.   
  }
}
