% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.GCElm.R
\name{summary.GCElm}
\alias{summary.GCElm}
\title{Summary method for GCElm objects}
\usage{
\method{summary}{GCElm}(object, correlation = FALSE, symbolic.cor = FALSE,
  ...)
}
\arguments{
\item{object}{An object of class \code{"GCElm"} returned by \code{GCElm}.}

\item{correlation}{logical; if TRUE, the correlation matrix of the estimated parameters is returned and printed.}

\item{symbolic.cor}{logical; If TRUE, print the correlations in a symbolic form (see symnum) rather than as numbers.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A \code{"summary.GCElm"} object with components described below:
\describe{
  \item{\code{call}}{The matched call.}
  \item{\code{coefficients}}{A matrix with estimates, standard errors, t values, and p-values.}
}
}
\description{
Provides a summary of a model fitted with \code{GCElm}, including estimated
coefficients, standard errors, test statistics, and p-values.
}
\examples{
\donttest{
set.seed(1234)
N <- 1000; K <- 10
betas <- c(-1, 1, -1, 1, rep(0, K-4))
X <- matrix(runif(N*K), nrow = N, ncol = K)
y <- as.vector(X \%*\% betas + rnorm(N))
Z <- matrix(rep(seq(-10, 10, 2.5), K + 1), nrow = K + 1, byrow = TRUE)
df <- data.frame(y, X)
fit <- GCElm(y ~ ., data = df, Z = Z)
}

}
