% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCElm.fit.r
\name{GCElm.fit}
\alias{GCElm.fit}
\title{Generalized Cross Entropy Linear Regression Models}
\usage{
GCElm.fit(y, X, Z, v, nu, p0, w0, k.sigma = 3, control = GCElm.control())
}
\arguments{
\item{y}{numeric, n (Nx1) vector representing the dependent variable where N is the number of observations.}

\item{X}{numeric, n (NxK) matrix representing a set of independent variables where K is the number of regressors.}

\item{Z}{numeric, an (KxM) matrix representing support spaces for the regression coefficients (including intercept) where M is the dimension of the support spaces.}

\item{v}{numeric, an optional argument representing a support space for error terms: 
\describe{
 \item{(a)}{if missing then \code{v} is a (5x1) vector of equally spaced points in [a,b] interval;}
 \item{(b)}{if a scalar (e.g. H) then \code{v} is a (Hx1) vector of equally spaced points in [a,b] interval;}
 \item{(c)}{can be a user-supplied vector;}
 \item{(d)}{can be a user-supplied matrix.}
}
Please note that in case (a) and (b) the [a,b] interval is centered around zero, and a and b are calculated using the empirical three-sigma rule Pukelsheim (1994).}

\item{nu}{numeric, an optional weight parameter representing the trade-off between prediction and precision.}

\item{p0}{numeric, optional prior probabilities associated with the regression coefficients.}

\item{w0}{numeric, optional prior probabilities associated with the error terms.}

\item{k.sigma}{numeric, coefficient \code{k} in the k-sigma rule (default \code{k=3}).}

\item{control}{list, a list of parameters for controlling the fitting process; for \code{GCElm.fit} this is passed to \code{\link{GCElm.control}}.}
}
\value{
A \code{list} with the following elements:
\itemize{
 \item{\code{lambda}, estimated lagrange multipliers;}
 \item{\code{coefficients}, regression coefficients;}
 \item{\code{var_beta}, variance-covariance matrix of the regression coefficients;}
 \item{\code{p}, estimated probabilities associated with the regressions coefficients;}
 \item{\code{w}, estimated probabilities associated with the error terms;}
 \item{\code{e}, estimated residuals;}
 \item{\code{Sp}, the (signal) information of the whole system;}
 \item{\code{Sp_k}, the (signal) information associated with the k-th regression coefficient;}
 \item{\code{H_p_w}, value of the joint entropies of p and w at the final iteration;}
 \item{\code{dH}, delta-H from the Entropy Concentration Theorem;}
 \item{\code{ER}, entropy-ratio statistic;}
 \item{\code{Pseudo-R2}, pseudo R-squared;}
 \item{\code{converged}, convergence (same as in the \code{\link[lbfgs]{lbfgs}} function).}
}
}
\description{
Fitting generalized cross entropy (GCE) linear models
}
\details{
Mettere qui eventuali details.
}
\examples{
set.seed(1234)
N <- 1000
K <- 10
betas <- c(-1,1,-1,1,0,0,0,0,0,0)
X <- matrix(runif(N*K), nrow = N, ncol = K)
y <- X \%*\% betas + rnorm(N)
X <- cbind(rep(1, N), X)
Z <- matrix(rep(seq(-10,10,2.5), K+1), nrow = K+1, byrow = TRUE)
GCEfit <- GCElm.fit(y, X, Z)
coef(GCEfit)      
}
\references{
Golan (1996)
}
\author{
Marco Sandri, Enrico Ciavolino, Maurizio Carpita (\email{gcemodels@gmail.com})
}
